+++ src/Bootstrapper/Cake.Bootstrapper/InstallCmdlet.cs
@@ -12,25 +12,29 @@ namespace Cake.Bootstrapper
     {
         [Parameter]
         public string Source { get; set; }
-
         [Parameter]
         public SwitchParameter AppVeyor { get; set; }
-
         [Parameter]
         public SwitchParameter GitIgnore { get; set; }
+        [Parameter]
+        public SwitchParameter Empty { get; set; }
 
         public override void RegisterDependencies(ContainerBuilder builder)
         {
+            // Register NuGet package configuration creator.
             builder.RegisterType<NuGetPackageConfigurationCreator>()
                 .As<INuGetPackageConfigurationCreator>()
                 .SingleInstance();
             
+            // Register NuGet version prober.
             builder.RegisterType<NuGetPackageVersionProber>()
                 .As<INugetPackageVersionProber>()
                 .SingleInstance();
 
+            // Register file copier.
             builder.RegisterType<FileCopier>().As<IFileCopier>().SingleInstance();
 
+            // Register gitignore patcher.
             builder.RegisterType<GitIgnorePatcher>().As<IGitIgnorePatcher>().SingleInstance();
         }
 
@@ -41,17 +45,20 @@ namespace Cake.Bootstrapper
             {
                 command.Source = Source;
             }
-
             // AppVeyor
             if (AppVeyor.IsPresent)
             {
                 command.AppVeyor = true;
             }
-
             // GitIgnore
             if (GitIgnore.IsPresent)
             {
                 command.GitIgnore = true;
+            }
+            // Empty script
+            if (Empty.IsPresent)
+            {
+                command.Empty = true;
             }
         }
     }