+++ src/Bootstrapper/Cake.Bootstrapper/Installer/Resources/IFileCopier.cs
@@ -2,6 +2,9 @@ namespace Cake.Bootstrapper.Installer.Resources
 {
     public interface IFileCopier
     {
-        bool Copy(string scriptName);
+        bool CopyConventionBasedCakeScript();
+        bool CopyEmptyCakeScript();
+        bool CopyAppVeyorConfiguration();
+        bool CopyBootstrapperScript();
     }
 }