+++ ReleaseNotes.md
@@ -1,3 +1,7 @@
+### New in 0.0.6
+* Added support for nightly build of Roslyn.
+* Added option to install empty cake script.
+
 ### New in 0.0.5 (Released 2015/01/17)
 * Added support for missing tools and package.config.
 * Bootstrapper script will now download Nuget if it doesn't exist.